+++ app/src/main/java/io/legado/app/ui/main/bookshelf/style1/books/BooksAdapterGrid.kt
@@ -42,7 +42,7 @@ class BooksAdapterGrid(context: Context, private val callBack: CallBack) :
             if (background is android.graphics.drawable.GradientDrawable) {
                 background.setColor(cardColor)
                 val dividerColor = ThemeStore.dividerColor(context)
-                background.setStroke((AppConfig.bookshelfCardBorderWidth * 0.5f).dpToPx(), dividerColor)
+                background.setStroke((AppConfig.bookshelfCardBorderWidth * 0.5f).dpToPx().toInt(), dividerColor)
             }
             holder.itemView.setPadding(8, 8, 8, 8)
         } else {
