+++ app/src/main/java/io/legado/app/ui/main/bookshelf/style1/books/BooksAdapterList2.kt
@@ -64,7 +64,7 @@ class BooksAdapterList2(
             if (background is android.graphics.drawable.GradientDrawable) {
                 background.setColor(cardColor)
                 val dividerColor = ThemeStore.dividerColor(context)
-                background.setStroke((AppConfig.bookshelfCardBorderWidth * 0.5f).dpToPx(), dividerColor)
+                background.setStroke((AppConfig.bookshelfCardBorderWidth * 0.5f).dpToPx().toInt(), dividerColor)
             }
             
             // 设置卡片左右边距为16dp，让卡片距离屏幕边缘16dp
