+++ app/src/main/java/io/legado/app/ui/main/bookshelf/style2/BooksAdapterList.kt
@@ -107,7 +107,7 @@ class BooksAdapterList(context: Context, callBack: CallBack) :
                 if (background is android.graphics.drawable.GradientDrawable) {
                     background.setColor(cardColor)
                     val dividerColor = ThemeStore.dividerColor(context)
-                    background.setStroke((AppConfig.bookshelfCardBorderWidth * 0.5f).dpToPx(), dividerColor)
+                    background.setStroke((AppConfig.bookshelfCardBorderWidth * 0.5f).dpToPx().toInt(), dividerColor)
                 }
                 
                 // 设置卡片左右边距为16dp，让卡片距离屏幕边缘16dp
@@ -440,7 +440,7 @@ class BooksAdapterList(context: Context, callBack: CallBack) :
                 if (background is android.graphics.drawable.GradientDrawable) {
                     background.setColor(cardColor)
                     val dividerColor = ThemeStore.dividerColor(context)
-                    background.setStroke((AppConfig.bookshelfCardBorderWidth * 0.5f).dpToPx(), dividerColor)
+                    background.setStroke((AppConfig.bookshelfCardBorderWidth * 0.5f).dpToPx().toInt(), dividerColor)
                 }
                 
                 // 设置卡片左右边距为16dp，让卡片距离屏幕边缘16dp
@@ -606,7 +606,7 @@ class BooksAdapterList(context: Context, callBack: CallBack) :
                 if (background is android.graphics.drawable.GradientDrawable) {
                     background.setColor(cardColor)
                     val dividerColor = ThemeStore.dividerColor(context)
-                    background.setStroke((AppConfig.bookshelfCardBorderWidth * 0.5f).dpToPx(), dividerColor)
+                    background.setStroke((AppConfig.bookshelfCardBorderWidth * 0.5f).dpToPx().toInt(), dividerColor)
                 }
                 
                 // 设置卡片左右边距为16dp，让卡片距离屏幕边缘16dp
